add new abstraction around providing PerformanceCounters in CounterBusiness
