+++ Tharga.Influx-Capacitor.Collector.Tests/Business/CounterBusinessTests/FilterTests.cs
@@ -141,7 +141,7 @@ namespace Tharga.InfluxCapacitor.Collector.Tests.Business.CounterBusinessTests
             var counters = Mocks.Of<ICounter>(z => z.CategoryName == "Processor" && z.CounterName == "% Processor Time" && z.InstanceName == "*").Take(1).ToList();
             var counterGroup = Mocks.Of<ICounterGroup>(y => y.Name == "A" && y.SecondsInterval == 10 && y.Counters == counters).Take(1).ToList();
             var config = Mock.Of<IConfig>(x => x.Groups == counterGroup);
-            var counterBusiness = new CounterBusiness(counter => new[]
+            var counterBusiness = new PerformanceCounterProvider(null, counter => new[]
             {
                 new PerformanceCounterInfo(null, null, "w3wp", null, null, null, null),
                 new PerformanceCounterInfo(null, null, "w3wp", null, null, null, null),
@@ -150,10 +150,10 @@ namespace Tharga.InfluxCapacitor.Collector.Tests.Business.CounterBusinessTests
             });
 
             // act
-            var result = counterBusiness.GetPerformanceCounterGroups(config).ToList();
+            var result = new[] { counterBusiness.GetGroup(counterGroup.First()) };
 
             // assert
-            Assert.That(result.Count, Is.EqualTo(1));
+            Assert.That(result.Length, Is.EqualTo(1));
             var resultCounters = result.First().GetCounters().ToList();
             Assert.That(resultCounters.Count, Is.EqualTo(4));
             Assert.That(resultCounters.First().InstanceName, Is.EqualTo("w3wp"));
