+++ Tharga.Influx-Capacitor.Collector/Tharga.Influx-Capacitor.Collector.csproj
@@ -94,6 +94,7 @@
     <Compile Include="Business\KafkaDataSender.cs" />
     <Compile Include="Business\MetaDataBusinessBusiness.cs" />
     <Compile Include="Business\NullDataSender.cs" />
+    <Compile Include="Business\PerformanceCounterProvider.cs" />
     <Compile Include="Business\PublisherBusiness.cs" />
     <Compile Include="Entities\CounterInstanceFilter.cs" />
     <Compile Include="Extensions\OutputLevelExtension.cs" />
@@ -108,6 +109,7 @@
     <Compile Include="Handlers\PublisherEngine.cs" />
     <Compile Include="Interface\ICounterInstanceFilter.cs" />
     <Compile Include="Interface\ICounterPublisher.cs" />
+    <Compile Include="Interface\IPerformanceCounterProvider.cs" />
     <Compile Include="Interface\IPublisherBusiness.cs" />
     <Compile Include="Publishers\Random.cs" />
     <Compile Include="Publishers\TotalMemory.cs" />
